@@ -1,19 +1,23 @@
 # -*- coding: utf-8 -*-
 # Much of the code here was forked from https://github.com/codelucas/newspaper
 # Copyright (c) Lucas Ou-Yang (codelucas)
+"""Module providing the Article class for newspaper. The Article class
+abstracts the concept of a news article, providing methods and properties
+to download, parse and analyze said article.
+"""
 
-
+from datetime import datetime
+import json
 import logging
-import copy
-import os
-import glob
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict, List, Literal, Optional, Set, Union, overload
 from urllib.parse import urlparse
 import lxml
 
 import requests
 
+from newspaper.exceptions import ArticleException
+from newspaper.text import StopWords
+import newspaper.parsers as parsers
 from . import network
 from . import nlp
 from . import settings
@@ -24,9 +28,6 @@ from .configuration import Configuration
 from .extractors import ContentExtractor
 from .outputformatters import OutputFormatter
 from .utils import (
-    URLHelper,
-    RawHelper,
-    extend_config,
     get_available_languages,
     extract_meta_refresh,
 )
@@ -51,12 +52,6 @@ class ArticleDownloadState:
     NOT_STARTED = 0
     FAILED_RESPONSE = 1
     SUCCESS = 2
-
-
-class ArticleException(Exception):
-    """Article Exception thrown by the article package."""
-
-    pass
 
 
 class Article:
@@ -90,21 +85,27 @@ class Article:
         images (List[str]): List of all image urls in the current article
         movies (List[str]): List of video links in the article body
         text (str): a parsed version of the article body. It will be truncated
-            to the first config.MAX_TEXT characters.
+            to the first `config.max_text` characters.
         text_cleaned (str): a parsed version of the clean_top_node content.
-            It will be truncated to the first config.MAX_TEXT characters.
+            It will be truncated to the first `config.max_text` characters.
+            .. deprecated:: 0.9.3
+                is now same as :any:`Article.text`
+                clean_top_node is removed
         keywords (List[str]): An inferred list of keywords for this article.
             This will be generated by the nlp method. It will be truncated to
-            the first config.MAX_KEYWORDS keywords.
+            the first `config.max_keywords` keywords.
+        keyword_scores (Dict[str, float]): A dictionary of keywords and their
+            scores.
         meta_keywords (List[str]):  A list of keywords provided by the meta data.
-            It will be truncated to the first config.MAX_KEYWORDS keywords.
+            It will be truncated to the first `config.max_keywords` keywords.
         tags (Set[str]): Extracted tag list from the article body
         authors (List[str]): The author list parsed from the article. It will
-            be truncated to the first config.MAX_AUTHORS authors.
+            be truncated to the first `config.max_authors` authors.
         publish_date (str): The parsed publishing date from the article. If no
             valid date is found, it will be an empty string.
         summary (str): The summarization of the article as generated by the nlp
-            method. It will be truncated to the first config.MAX_SUMMARY_SENT sentences.
+            method. It will be truncated to the first `config.max_summary_sent`
+            sentences.
         html (str): The raw html of the article page.
         article_html (str): The raw html of the article body.
         is_parsed (bool): True if parse() has been called.
@@ -112,6 +113,7 @@ class Article:
             successful, ArticleDownloadState.FAILED_RESPONSE if `download()` failed,
             `ArticleDownloadState.NOT_STARTED` if `download()` was not called.
         download_exception_msg (str): The exception message if download() failed.
+        history (List[str]): Redirection history from the requests.get call.
         meta_description (str): The description extracted from the meta data.
         meta_lang (str): The language extracted from the meta data.
             If config.language is not set, this value will be used
@@ -124,15 +126,12 @@ class Article:
         top_node (lxml.html.HtmlElement): Top node of the original DOM tree.
             It contains the text nodes for the detected article body. This node
             is on the doc DOM tree.
-        clean_top_node (lxml.html.HtmlElement): Top node for the article on the
-            cleaned version of the DOM. This node is _not_ in the doc DOM tree.
+
         doc (lxml.html.HtmlElement): the full DOM of the downloaded html. It is
             the original DOM tree.
         clean_doc (lxml.html.HtmlElement): a cleaned version of the DOM tree
-        additional_data (Dict[Any, Any]): A property dict for users to store
-            custom data.
-        link_hash (str): a unique hash for the url of this article. It is salted
-            with the timestamp of the download.
+            .. deprecated:: 0.9.3
+                is now same as :any:`Article.doc`
     """
 
     def __init__(
@@ -142,19 +141,23 @@ class Article:
         source_url: str = "",
         read_more_link: str = "",
         config: Optional[Configuration] = None,
-        **kwargs: Dict[str, Any],
+        **kwargs: Any,
     ):
-        """Constructs the article class. Will not download or parse the article
+        """
+        Constructs the article class. Will not download or parse the article
 
         Args:
+
             url (str): The input url to parse. Can be a URL or a file path.
             title (str, optional): Default title if none can be
-                extracted from the webpage. Defaults to "".
+                extracted from the webpage. Defaults to empty string.
             source_url (str, optional): URL of the main website that
                 originates the article.
                 If left empty, it will be inferred from the url. Defaults to "".
             read_more_link (str, optional): A xpath selector for the link to the
-                full article. make sure that the selector works for all casese,
+                full article, in case there is a 'preview' with a read-more
+                button that leads to another url (even on another domain).
+                make sure that the selector works for all cases,
                 not only for one specific article. If needed, you can use
                 several xpath selectors separated by `|`. Defaults to "".
             config (Configuration, optional): Configuration settings for
@@ -162,8 +165,8 @@ class Article:
             use the default settingsDefaults to None.
 
         Keyword Args:
-            **kwargs: Any Configuration class propriety can be overwritten
-                    through init keyword  params.
+            **kwargs: Any Configuration class property can be overwritten
+                    through init keyword params.
                     Additionally, you can specify any of the following
                     requests parameters:
                     headers, cookies, auth, timeout, allow_redirects,
@@ -184,7 +187,7 @@ class Article:
             if k in kwargs:
                 self.config.requests_params[k] = kwargs[k]
                 del kwargs[k]
-        self.config = extend_config(self.config, kwargs)
+        self.config.update(**kwargs)
 
         self.extractor = ContentExtractor(self.config)
 
@@ -192,7 +195,7 @@ class Article:
             scheme = urls.get_scheme(url)
             if scheme is None:
                 scheme = "http"
-            source_url = scheme + "://" + urls.get_domain(url)
+            source_url = scheme + "://" + str(urls.get_domain(url))
 
         if source_url is None or source_url == "":
             raise ArticleException("input url bad format")
@@ -205,6 +208,7 @@ class Article:
         # In case of follow read more link, we need to keep the original url
         self.original_url = self.url
 
+        self._title = title
         self.title = title
 
         # An xpath that allows to find the link to the full article
@@ -223,11 +227,13 @@ class Article:
         self.movies: List[str] = []
 
         # Body text from this article
-        self.text = ""
-        self.text_cleaned = ""
+        self._text = ""
 
         # `keywords` are extracted via nlp() from the body text
         self.keywords: List[str] = []
+
+        # `keyword_scores` a dictionary of keywords and their scores
+        self.keyword_scores: Dict[str, float] = {}
 
         # `meta_keywords` are extracted via parse() from <meta> tags
         self.meta_keywords: List[str] = []
@@ -238,13 +244,13 @@ class Article:
         # List of authors who have published the article, via parse()
         self.authors: List[str] = []
 
-        self.publish_date = ""
+        self.publish_date: Optional[datetime] = None
 
         # Summary generated from the article's body txt
-        self.summary = ""
+        self._summary = ""
 
         # This article's unchanged and raw HTML
-        self.html = ""
+        self._html = ""
 
         # The HTML of this article's main node (most important part)
         self.article_html = ""
@@ -253,6 +259,9 @@ class Article:
         self.is_parsed = False
         self.download_state = ArticleDownloadState.NOT_STARTED
         self.download_exception_msg: Optional[str] = None
+
+        # Redirection history from the requests.get call
+        self.history: Optional[List[str]] = []
 
         # Meta description field in the HTML source
         self.meta_description = ""
@@ -276,25 +285,12 @@ class Article:
         # for the main body of the article
         self.top_node: Optional[lxml.html.Element] = None
 
-        # A deepcopied clone of the above object before heavy parsing
-        # operations, useful for users to query data in the
-        # "most important part of the page"
-        self.clean_top_node: Optional[lxml.html.Element] = None
-
         # The top node complemented with siblings (off-tree)
         self._top_node_complemented: Optional[lxml.html.Element] = None
 
         # lxml DOM object generated from HTML
         self.doc: Optional[lxml.html.Element] = None
-
-        # A deepcopied clone of the above object before undergoing heavy
-        # cleaning operations, serves as an API if users need to query the DOM
-        self.clean_doc: Optional[lxml.html.Element] = None
-
-        # A property dict for users to store custom data.
-        self.additional_data: Dict[Any, Any] = {}
-
-        self.link_hash: Optional[str] = None
+        self._clean_doc: Optional[lxml.html.Element] = None
 
     def build(self):
         """Build a lone article from a URL independent of the source (newspaper).
@@ -317,8 +313,13 @@ class Article:
 
     def _parse_scheme_http(self, url: Optional[str] = None):
         try:
-            html, stauts_code = network.get_html_2XX_only(url or self.url, self.config)
-            if stauts_code >= 400:
+            # We do not use get_html() here because we want to be able to
+            # detect protection in the response regardless of the status code
+            html, status_code, history = network.get_html_status(
+                url or self.url, self.config
+            )
+            self.history = [r.url for r in history]
+            if status_code >= 400:
                 self.download_state = ArticleDownloadState.FAILED_RESPONSE
                 protection = self._detect_protection(html)
                 if protection:
@@ -327,7 +328,7 @@ class Article:
                     )
                 else:
                     self.download_exception_msg = (
-                        f"Status code {stauts_code} for url {url}"
+                        f"Status code {status_code} for url {url}"
                     )
                 return None
         except requests.exceptions.RequestException as e:
@@ -342,6 +343,8 @@ class Article:
             return "Cloudflare"
         if "/cdn-cgi/challenge-platform/h/b/orchestrate/chl_page" in html:
             return "Cloudflare"
+        if "cloud-flare" in html:
+            return "Cloudflare"
         if "CloudFront" in html:
             return "CloudFront"
         if "perimeterx" in html:
@@ -355,7 +358,7 @@ class Article:
         title: Optional[str] = None,
         recursion_counter: int = 0,
         ignore_read_more: bool = False,
-    ):
+    ) -> "Article":
         """Downloads the link's HTML content, don't use if you are batch async
         downloading articles
 
@@ -372,6 +375,8 @@ class Article:
             ignore_read_more (bool, optional): If true, the download process will
             ignore any kind of "read_more" xpath set up in the constructor.
             Defaults to False.
+        Returns:
+            Article: self
         """
 
         if input_html is None:
@@ -386,7 +391,7 @@ class Article:
                     self.url,
                     self.download_exception_msg,
                 )
-                return
+                return self
         else:
             html = input_html
 
@@ -397,8 +402,9 @@ class Article:
                     input_html=network.get_html(meta_refresh_url),
                     recursion_counter=recursion_counter + 1,
                 )
+
         if not ignore_read_more and self.read_more_link:
-            doc = self.config.get_parser().fromstring(html)
+            doc = parsers.fromstring(html)
             for read_more_node in doc.xpath(self.read_more_link):
                 # TODO: add check for onclick redirections. need some examples
                 if read_more_node.get("href"):
@@ -426,47 +432,45 @@ class Article:
                         )
                     break
 
-        self.set_html(html)
-        self.set_title(title)
+        self.html = html
+        if title is not None:
+            self.title = title
 
-    def parse(self):
+        return self
+
+    def parse(self) -> "Article":
         """Parse the previously downloaded article.
         If `download()` wasn't called, it will raise
         a `ArticleException` exception.
         Populates the article properties such as:
         ``title``, ``authors``, ``publish_date``,
         ``text``, ``top_image``, etc.
-
+        Returns:
+            Article: self
         """
         self.throw_if_not_downloaded_verbose()
 
-        self.doc = self.config.get_parser().fromstring(self.html)
-        self.clean_doc = copy.deepcopy(self.doc)
+        self.doc = parsers.fromstring(self.html)
 
         if self.doc is None:
             # `parse` call failed, return nothing
-            return
-
-        # TODO: Fix this, sync in our fix_url() method
-        parse_candidate = self.get_parse_candidate()
-        self.link_hash = parse_candidate.link_hash  # MD5
+            self.is_parsed = True
+            return self
 
         document_cleaner = DocumentCleaner(self.config)
         output_formatter = OutputFormatter(self.config)
 
-        title = self.extractor.get_title(self.clean_doc)
-        self.set_title(title)
+        title = self.extractor.get_title(self.doc)
+        self.title = title
 
-        authors = self.extractor.get_authors(self.clean_doc)
-        self.set_authors(authors)
+        authors = self.extractor.get_authors(self.doc)
+        self.authors = authors[: self.config.max_authors]
 
-        metadata = self.extractor.get_metadata(self.url, self.clean_doc)
+        metadata = self.extractor.get_metadata(self.url, self.doc)
         if metadata["language"] in get_available_languages():
             self.meta_lang = metadata["language"]
-
             if self.config.use_meta_language:
-                self.extractor.update_language(self.meta_lang)
-                output_formatter.update_language(self.meta_lang)
+                self.config.language = metadata["language"]
 
         self.meta_site_name = metadata["site_name"]
         self.meta_description = metadata["description"]
@@ -475,45 +479,37 @@ class Article:
         self.tags = metadata["tags"]
         self.meta_data = metadata["data"]
 
-        self.publish_date = self.extractor.get_publishing_date(self.url, self.clean_doc)
-
-        # Before any computations on the body, clean DOM object
-        self.clean_doc = document_cleaner.clean(self.clean_doc)
+        self.publish_date = self.extractor.get_publishing_date(self.url, self.doc)
 
         # Top node in the original documentDOM
         self.top_node = self.extractor.calculate_best_node(self.doc)
-
-        # Top node in the cleaned version of the DOM
-        self.clean_top_node = self.extractor.calculate_best_node(self.clean_doc)
+        # Off-tree Node containing the top node and any relevant siblings
+        self._top_node_complemented = self.extractor.top_node_complemented
 
         self.set_movies(self.extractor.get_videos(self.doc, self.top_node))
 
-        if self.top_node is not None:
-            # Off-tree Node containing the top node and any relevant siblings
-            self._top_node_complemented = self.extractor.top_node_complemented
-
-            text, article_html = output_formatter.get_formatted(
-                self._top_node_complemented
-            )
-            self.set_article_html(article_html)
-            self.set_text(text)
-
-            text, _ = output_formatter.get_formatted(self.clean_top_node)
-            self.text_cleaned = text[: self.config.MAX_TEXT] if text else ""
-
         self.fetch_images()
 
+        if self.top_node is not None:
+            self._top_node_complemented = document_cleaner.clean(
+                self._top_node_complemented
+            )
+            text, article_html = output_formatter.get_formatted(
+                self._top_node_complemented, title
+            )
+            self.article_html = article_html
+            self.text = text
+
         self.is_parsed = True
-        self.release_resources()
+        return self
 
     def fetch_images(self):
         """Fetch top image, meta image and image list from
         current cleaned_doc. Will set the attributes: meta_img,
         top_image, images, meta_favicon
         """
-        # TODO: check weather doc or clean doc is better
         # TODO: rewrite set_reddit_top_img. I removed it for now
-        self.extractor.parse_images(self.url, self.clean_doc, self.clean_top_node)
+        self.extractor.parse_images(self.url, self.doc, self.top_node)
 
         self.meta_img = self.extractor.image_extractor.meta_image
         self.top_image = self.extractor.image_extractor.top_image
@@ -539,7 +535,7 @@ class Article:
         wordcount = self.text.split(" ")
         sentcount = self.text.split(".")
 
-        if meta_type == "article" and len(wordcount) > (self.config.MIN_WORD_COUNT):
+        if meta_type == "article" and len(wordcount) > (self.config.min_word_count):
             log.debug("%s verified for article and wc", self.url)
             return True
 
@@ -551,11 +547,11 @@ class Article:
             log.debug("%s caught for bad title", self.url)
             return False
 
-        if len(wordcount) < self.config.MIN_WORD_COUNT:
+        if len(wordcount) < self.config.min_word_count:
             log.debug("%s caught for word cnt", self.url)
             return False
 
-        if len(sentcount) < self.config.MIN_SENT_COUNT:
+        if len(sentcount) < self.config.min_sent_count:
             log.debug("%s caught for sent cnt", self.url)
             return False
 
@@ -590,79 +586,89 @@ class Article:
         self.throw_if_not_downloaded_verbose()
         self.throw_if_not_parsed_verbose()
 
-        nlp.load_stopwords(self.config.language)
-        text_keyws = list(nlp.keywords(self.text).keys())
-        title_keyws = list(nlp.keywords(self.title).keys())
-        keyws = list(set(title_keyws + text_keyws))
-        self.set_keywords(keyws)
+        stopwords = StopWords(self.config.language)
+        keywords = nlp.keywords(self.text, stopwords, self.config.max_keywords)
+        for k, v in nlp.keywords(
+            self.title, stopwords, self.config.max_keywords
+        ).items():
+            if k in keywords:
+                keywords[k] += v
+                keywords[k] /= 2
+            else:
+                keywords[k] = v
 
-        max_sents = self.config.MAX_SUMMARY_SENT
+        keywords = sorted(list(keywords.items()), key=lambda x: x[1], reverse=True)
+        keywords = keywords[: self.config.max_keywords]
+
+        self.keywords = [x[0] for x in keywords]  # remove score
+        self.keyword_scores = dict(keywords)
+
+        max_sents = self.config.max_summary_sent
 
         summary_sents = nlp.summarize(
-            title=self.title, text=self.text, max_sents=max_sents
+            title=self.title, text=self.text, stopwords=stopwords, max_sents=max_sents
         )
-        summary = "\n".join(summary_sents)
-        self.set_summary(summary)
+        self.summary = "\n".join(summary_sents)
 
-    def get_parse_candidate(self):
-        """A parse candidate is a wrapper object holding a link hash of this
-        article and a final_url of the article
+    @property
+    def title(self) -> str:
+        """Get the title of the article.
+        Returns:
+            str: The title of the article.
         """
-        if self.html:
-            return RawHelper.get_parsing_candidate(self.url, self.html)
-        return URLHelper.get_parsing_candidate(self.url)
+        return self._title
 
-    def build_resource_path(self):
-        """Must be called after computing HTML/final URL"""
-        res_path = Path(self.get_resource_path())
-        if not res_path.exists():
-            res_path.mkdir(parents=True, exist_ok=True)
-
-    def get_resource_path(self):
-        """Every article object has a special directory to store data in from
-        initialization to garbage collection
+    @title.setter
+    def title(self, value: str):
+        """Set the title of the article.
+        Args:
+            value (str): The title of the article.
         """
-        resource_directory = Path(settings.TOP_DIRECTORY) / "article_resources"
-        resource_directory.mkdir(parents=True, exist_ok=True)
-        dir_path = resource_directory / f"{self.link_hash}_"
-        return str(dir_path)
+        self._title = value[: self.config.max_title] if value else ""
 
-    def release_resources(self):
-        # TODO: implement in entirety
-        path = self.get_resource_path()
-        for fname in glob.glob(path):
-            try:
-                os.remove(fname)
-            except OSError:
-                pass
-        # os.remove(path)
+    @property
+    def text(self) -> str:
+        """Returns the text content of the article.
+        Returns:
+            The text content of the article.
+        """
+        return self._text
 
-    def set_title(self, input_title):
-        if input_title:
-            self.title = input_title[: self.config.MAX_TITLE]
+    @text.setter
+    def text(self, value: str):
+        """Sets the text of the article.
+        Args:
+            value (str): The text to be set.
+        """
+        self._text = value[: self.config.max_text] if value else ""
 
-    def set_text(self, text):
-        text = text[: self.config.MAX_TEXT]
-        if text:
-            self.text = text
+    @property
+    def html(self) -> str:
+        """Returns the HTML content of the article.
+        Returns:
+            The HTML content of the article.
+        """
+        return self._html
 
-    def set_html(self, html):
-        """Encode HTML before setting it"""
-        if html:
-            if isinstance(html, bytes):
-                html = self.config.get_parser().get_unicode_html(html)
-            self.html = html
-            self.download_state = ArticleDownloadState.SUCCESS
-
-    def set_article_html(self, article_html):
-        """Sets the HTML of just the article's `top_node`"""
-        if article_html:
-            self.article_html = article_html
+    @html.setter
+    def html(self, value: str):
+        """Sets the HTML content of the article.
+        Args:
+            value (str): The HTML content to set.
+        """
+        self.download_state = ArticleDownloadState.SUCCESS
+        if value:
+            if isinstance(value, bytes):
+                value = parsers.get_unicode_html(value)
+            self._html = value
+        else:
+            self._html = ""
 
     @property
     def imgs(self) -> List[str]:
-        """Same as images
-
+        """Same as images. Kept for legacy reasons
+        .. deprecated:: 0.9.3
+            use :any:`Article.images` instead
         Returns:
             List[str]: list of image urls
         """
@@ -671,36 +677,52 @@ class Article:
 
     @property
     def top_img(self) -> str:
-        """Same as top_image
-
+        """Same as top_image. Kept for legacy reasons
+        .. deprecated:: 0.9.3
+            use :any:`Article.top_image` instead
         Returns:
-            str: top_image
+            str: Top image of the article
         """
         # Seems to be some legacy api,
         return self.top_image
 
-    def set_keywords(self, keywords):
-        """Keys are stored in list format"""
-        if not isinstance(keywords, list):
-            raise ValueError("Keyword input must be list!")
-        if keywords:
-            self.keywords = keywords[: self.config.MAX_KEYWORDS]
-
-    def set_authors(self, authors):
-        """Authors are in ["firstName lastName", "firstName lastName"] format"""
-        if not isinstance(authors, list):
-            raise ValueError("authors input must be list!")
-        if authors:
-            self.authors = authors[: self.config.MAX_AUTHORS]
-
-    def set_summary(self, summary):
-        """Summary here refers to a paragraph of text from the
-        title text and body text
+    @property
+    def clean_doc(self) -> lxml.html.HtmlElement:
+        """Cleans the document and returns the cleaned version.
+        Returns:
+            lxml.html.HtmlElement: The cleaned document.
         """
-        self.summary = summary[: self.config.MAX_SUMMARY]
+        if self._clean_doc is None:
+            document_cleaner = DocumentCleaner(self.config)
+            self._clean_doc = document_cleaner.clean(self._clean_doc)
+        return self._clean_doc
+
+    @property
+    def text_cleaned(self) -> str:
+        """Returns the cleaned text of the article.
+        .. deprecated:: 0.9.3
+            use :any:`Article.text` instead. Text Cleaned does not make sense
+            anymore, as it was of lower quality than the regular text.
+        Returns:
+            str: The cleaned text of the article.
+        """
+        return self.text
+
+    @property
+    def summary(self) -> str:
+        """Returns the summary of the article."""
+        return self._summary
+
+    @summary.setter
+    def summary(self, value: str):
+        """Set the summary of the article.
+        Args:
+            value (str): The summary text.
+        """
+        self._summary = value[: self.config.max_summary] if value else ""
 
     def set_movies(self, movie_objects):
-        """Trim video objects into just urls"""
+        """Set the video urls from Video Objects"""
         movie_urls = [o.src for o in movie_objects if o and o.src]
         self.movies = movie_urls
 
@@ -722,3 +744,125 @@ class Article:
         """
         if not self.is_parsed:
             raise ArticleException("You must `parse()` an article first!")
+
+    @overload
+    def to_json(self, as_string: Literal[True]) -> str:
+        pass
+
+    @overload
+    def to_json(self, as_string: Literal[False]) -> Dict:
+        pass
+
+    def to_json(self, as_string: Optional[bool] = True) -> Union[str, Dict]:
+        """Create a json string from the article data. It will include the most
+        important attributes such as title, text, authors, publish_date, etc.
+        Must be called after `parse()`
+
+        Arguments:
+            as_string (bool, optional): If True, it will return a json string.
+                If False, it will return a json object. Defaults to True.
+
+        Returns:
+            str: the json string version of an parsed article.
+        """
+
+        self.throw_if_not_parsed_verbose()
+
+        article_dict: Dict[str, Any] = {}
+
+        for metadata in settings.article_json_fields:
+            value = getattr(self, metadata, getattr(self.config, metadata, None))
+            if isinstance(value, datetime):
+                article_dict[metadata] = value.isoformat()
+            else:
+                article_dict[metadata] = value
+        if as_string:
+            return json.dumps(article_dict, indent=4, ensure_ascii=False)
+        else:
+            return article_dict
+
+    def __getstate__(self):
+        """Return a pickable object for this article. This can be used for caching"""
+        state = self.__dict__.copy()
+        # drop non pickable attributes
+        if (
+            self.download_state == ArticleDownloadState.SUCCESS
+            and self.top_node is not None
+        ):
+            state["__parsed_state"] = True
+            self.top_node.set("__newspaper_top_node", "xxx")
+            self._top_node_complemented.set("__newspaper_top_node_complemented", "xxx")
+            state["_doc_html"] = parsers.node_to_string(self.doc)
+        else:
+            state["__parsed_state"] = False
+
+        state.pop("extractor", None)
+        state.pop("top_node", None)
+        state.pop("_top_node_complemented", None)
+        state.pop("doc", None)
+        # state.pop("clean_doc", None)
+
+        return state
+
+    def __setstate__(self, state):
+        """Restore state from the unpickled state"""
+        self.__dict__.update(state)
+        self.extractor = ContentExtractor(self.config)
+        self.top_node = None
+        self._top_node_complemented = None
+        self.doc = None
+        # self.clean_doc = None
+
+        if state["__parsed_state"]:
+            self.doc = parsers.fromstring(state["_doc_html"])
+            delattr(self, "_doc_html")
+            nodes = parsers.get_elements_by_attribs(
+                self.doc, attribs={"__newspaper_top_node": "xxx"}
+            )
+            if nodes:
+                self.top_node = nodes[0]
+            nodes = parsers.get_elements_by_attribs(
+                self.doc, attribs={"__newspaper_top_node_complemented": "xxx"}
+            )
+            if nodes:
+                self._top_node_complemented = nodes[0]
+        delattr(self, "__parsed_state")
+
+    def __eq__(self, other):
+        """Compare two Article objects. If they are the same, return True.
+        Otherwise, return False.
+        Args:
+            other (Any): Another object to compare to.
+        Returns:
+            bool: True if the objects are the same, False otherwise.
+        """
+        if not isinstance(other, Article):
+            raise NotImplementedError("Can only compare to other Article objects")
+
+        criteria = [
+            self.url == other.url,
+            self.title == other.title,
+            self.text == other.text,
+            self.top_image == other.top_image,
+            sorted(self.movies) == sorted(other.movies),
+            sorted(self.authors) == sorted(other.authors),
+            sorted(self.keywords) == sorted(other.keywords),
+            sorted(self.images) == sorted(other.images),
+            self.publish_date == other.publish_date,
+        ]
+
+        return all(criteria)
+
+    def __str__(self):
+        """Return a string representation of the article.
+        Returns:
+            str: A string representation of the article.
+        """
+        repr_ = f"__Title__: {self.title}"
+
+        if len(self.text) > 100:
+            repr_ += f"\n\n {self.text[:50]} [...] {self.text[:-50]}"
+        else:
+            repr_ += f"\n\n {self.text}"
+
+        return repr_
