@@ -6,17 +6,22 @@
 Holds the code for cleaning out unwanted tags from the lxml
 dom xpath.
 """
-import copy
-from .utils import ReplaceSequence
+import re
+from lxml.html import HtmlElement
+import newspaper.parsers as parsers
+from newspaper.configuration import Configuration
 
 
 class DocumentCleaner:
-    def __init__(self, config):
+    """
+    A class that provides methods to clean and manipulate HTML documents.
+    """
+
+    def __init__(self, config: Configuration):
         """Set appropriate tag names and regexes of tags to remove
         from the HTML
         """
         self.config = config
-        self.parser = self.config.get_parser()
         self.remove_nodes_re = (
             "^side$|combx|retweet|mediaarticlerelated|menucontainer|"
             "navbar|storytopbar-bucket|utility-bar|inline-share-tools"
@@ -31,27 +36,22 @@ class DocumentCleaner:
             "|konafilter|KonaFilter|breadcrumbs|^fn$|wp-caption-text"
             "|legende|ajoutVideo|timestamp|js_replies"
         )
-        self.regexp_namespace = "http://exslt.org/regular-expressions"
-        self.nauthy_ids_re = "//*[re:test(@id, '%s', 'i')]" % self.remove_nodes_re
-        self.nauthy_classes_re = (
-            "//*[re:test(@class, '%s', 'i')]" % self.remove_nodes_re
+        self.remove_nodes_related_re = (
+            r"related[-\s\_]?(search|topics|media|info|tags|article|content|links)|"
+            r"(search|topics|media|info|tags|article|content|links)[-\s\_]?related"
         )
-        self.nauthy_names_re = "//*[re:test(@name, '%s', 'i')]" % self.remove_nodes_re
         self.div_to_p_re = r"<(a|blockquote|dl|div|img|ol|p|pre|table|ul)"
         self.caption_re = "^caption$"
         self.google_re = " google "
         self.entries_re = "^[^entry-]more.*$"
         self.facebook_re = "[^-]facebook"
         self.facebook_broadcasting_re = "facebook-broadcasting"
-        self.twitter_re = "[^-]twitter"
-        self.tablines_replacements = (
-            ReplaceSequence().create("\n", "\n\n").append("\t").append("^\\s+$")
-        )
+        self.twitter_re = "[^-]twitter|twitter-tweet"
         self.contains_article = (
             './/article|.//*[@id="article"]|.//*[contains(@itemprop,"articleBody")]'
         )
 
-    def clean(self, doc_to_clean):
+    def clean(self, doc_to_clean: HtmlElement) -> HtmlElement:
         """Remove chunks of the DOM as specified"""
         doc_to_clean = self.clean_body_classes(doc_to_clean)
         doc_to_clean = self.clean_article_tags(doc_to_clean)
@@ -59,185 +59,257 @@ class DocumentCleaner:
         doc_to_clean = self.remove_drop_caps(doc_to_clean)
         doc_to_clean = self.remove_scripts_styles(doc_to_clean)
         doc_to_clean = self.clean_bad_tags(doc_to_clean)
+
+        # Remove image captions
         doc_to_clean = self.remove_nodes_regex(doc_to_clean, self.caption_re)
+        doc_to_clean = self.clean_caption_tags(doc_to_clean)
+
         doc_to_clean = self.remove_nodes_regex(doc_to_clean, self.google_re)
         doc_to_clean = self.remove_nodes_regex(doc_to_clean, self.entries_re)
+
+        # Remove social media cards
         doc_to_clean = self.remove_nodes_regex(doc_to_clean, self.facebook_re)
+        doc_to_clean = self.remove_nodes_regex(doc_to_clean, self.twitter_re)
         doc_to_clean = self.remove_nodes_regex(
             doc_to_clean, self.facebook_broadcasting_re
         )
-        doc_to_clean = self.remove_nodes_regex(doc_to_clean, self.twitter_re)
+        # Remove "related" sections
+        doc_to_clean = self.remove_nodes_regex(
+            doc_to_clean, self.remove_nodes_related_re
+        )
+
+        # Remove spans inside of paragraphs
         doc_to_clean = self.clean_para_spans(doc_to_clean)
-        doc_to_clean = self.div_to_para(doc_to_clean, "div")
-        doc_to_clean = self.div_to_para(doc_to_clean, "span")
-        doc_to_clean = self.div_to_para(doc_to_clean, "section")
+
+        doc_to_clean = self.reduce_article(doc_to_clean)
+
         return doc_to_clean
 
-    def clean_body_classes(self, doc):
+    def clean_whitespace(self, text: str) -> str:
+        """Remove tabs, whitespace lines from text
+        add double newlines to paragraphs
+        """
+        text = text.replace("\t", "")
+        text = re.sub(r"(?<!\n)\n(?!\n)", "\n\n", text)
+        text = re.sub(r"^\s+$", "", text)
+        return text
+
+    def clean_body_classes(self, doc: HtmlElement) -> HtmlElement:
         """Removes the `class` attribute from the <body> tag because
         if there is a bad match, the entire DOM will be empty!
         """
-        elements = self.parser.getElementsByTag(doc, tag="body")
+        elements = parsers.get_tags(doc, tag="body")
         if elements:
-            self.parser.delAttribute(elements[0], attr="class")
+            # Remove attribute
+            elements[0].attrib.pop("class", None)
         return doc
 
-    def clean_article_tags(self, doc):
-        articles = self.parser.getElementsByTag(doc, tag="article")
+    def clean_article_tags(self, doc: HtmlElement) -> HtmlElement:
+        """Removes specified attributes from <article> tags in the given document.
+        Args:
+            doc (ElementTree.Element): The document to clean.
+        Returns:
+            ElementTree.Element: The cleaned document.
+        """
+        articles = parsers.get_tags(doc, tag="article")
+
+        # Remove specified attributes from every <article> tag
         for article in articles:
             for attr in ["id", "name", "class"]:
-                self.parser.delAttribute(article, attr=attr)
+                article.attrib.pop(attr, None)
         return doc
 
-    def clean_em_tags(self, doc):
-        ems = self.parser.getElementsByTag(doc, tag="em")
+    def clean_em_tags(self, doc: HtmlElement) -> HtmlElement:
+        """Removes <em> tags from the given HTML document if they
+        don't contain any <img> tags.
+
+        Args:
+            doc (HtmlElement): The HTML document to clean.
+
+        Returns:
+            HtmlElement: The cleaned HTML document.
+        """
+        ems = parsers.get_tags(doc, tag="em")
         for node in ems:
-            images = self.parser.getElementsByTag(node, tag="img")
+            images = parsers.get_tags(node, tag="img")
             if len(images) == 0:
-                self.parser.drop_tag(node)
+                parsers.drop_tags(node)
         return doc
 
-    def remove_drop_caps(self, doc):
-        items = self.parser.css_select(
-            doc, "span[class~=dropcap], span[class~=drop_cap]"
+    def clean_caption_tags(self, doc: HtmlElement) -> HtmlElement:
+        """Removes image caption tags from the given HTML document.
+
+        Args:
+            doc (HtmlElement): The HTML document to clean.
+
+        Returns:
+            HtmlElement: The cleaned HTML document.
+        """
+        captions = parsers.get_tags(doc, tag="figure")
+        parsers.remove(captions, keep_tags=["img"])
+
+        captions = parsers.get_tags(doc, tag="figcaption")
+        parsers.remove(captions, keep_tags=["img"])
+
+        captions = parsers.get_tags(doc, attribs={"itemprop": "caption"})
+        parsers.remove(captions)
+
+        captions = parsers.get_tags(doc, attribs={"class": "instagram-media"})
+        parsers.remove(captions)
+
+        captions = parsers.get_tags(doc, attribs={"class": "image-caption"})
+        parsers.remove(captions)
+
+        captions = parsers.get_tags(
+            doc, attribs={"class": "caption"}, attribs_match="substring"
         )
-        for item in items:
-            self.parser.drop_tag(item)
+        captions = [c for c in captions if c.tag in ["div", "span", "header"]]
+        parsers.remove(captions)
+
         return doc
 
-    def remove_scripts_styles(self, doc):
+    def remove_drop_caps(self, doc: HtmlElement) -> HtmlElement:
+        """Removes spans with ckass dropcap from the given HTML document.
+
+        Args:
+            doc (HtmlElement): The HTML document to remove drop caps from.
+
+        Returns:
+            HtmlElement: The modified HTML document without drop caps.
+        """
+        items = parsers.get_tags(
+            doc, "span", {"class": "dropcap"}, attribs_match="word"
+        )
+        items += parsers.get_tags(
+            doc, "span", {"class": "drop_cap"}, attribs_match="word"
+        )
+
+        parsers.drop_tags(items)
+        return doc
+
+    def remove_scripts_styles(self, doc: HtmlElement) -> HtmlElement:
+        """Removes scripts, styles, and comments from the given HTML document.
+
+        Args:
+            doc (HtmlElement): The HTML document to remove scripts,
+            styles, and comments from.
+
+        Returns:
+            HtmlElement: The modified HTML document with scripts, styles,
+            and comments removed.
+        """
         # remove scripts
-        scripts = self.parser.getElementsByTag(doc, tag="script")
+        scripts = parsers.get_tags(doc, tag="script")
         for item in scripts:
-            self.parser.remove(item)
+            parsers.remove(item)
         # remove styles
-        styles = self.parser.getElementsByTag(doc, tag="style")
+        styles = parsers.get_tags(doc, tag="style")
         for item in styles:
-            self.parser.remove(item)
-        # remove comments
-        comments = self.parser.getComments(doc)
-        for item in comments:
-            self.parser.remove(item)
+            parsers.remove(item)
+        # remove comments <--! like this one -->
+        comments = doc.xpath("//comment()")
+
+        parsers.remove(comments)
 
         return doc
 
-    def clean_bad_tags(self, doc):
-        # ids
-        naughty_list = self.parser.xpath_re(doc, self.nauthy_ids_re)
+    def clean_bad_tags(self, doc: HtmlElement) -> HtmlElement:
+        """Cleans some known bad tags from the given HTML document.
+
+        Args:
+            doc (HtmlElement): The HTML document to clean.
+
+        Returns:
+            HtmlElement: The cleaned HTML document.
+        """
+        # bad ids
+        naughty_list = parsers.get_tags_regex(doc, attribs={"id": self.remove_nodes_re})
         for node in naughty_list:
             if not node.xpath(self.contains_article):
-                self.parser.remove(node)
+                parsers.remove(node)
         # class
-        naughty_classes = self.parser.xpath_re(doc, self.nauthy_classes_re)
-        for node in naughty_classes:
+        naughty_list = parsers.get_tags_regex(
+            doc, attribs={"class": self.remove_nodes_re}
+        )
+        for node in naughty_list:
             if not node.xpath(self.contains_article):
-                self.parser.remove(node)
+                parsers.remove(node)
         # name
-        naughty_names = self.parser.xpath_re(doc, self.nauthy_names_re)
-        for node in naughty_names:
-            if not node.xpath(self.contains_article):
-                self.parser.remove(node)
+        naughty_list = parsers.get_tags_regex(
+            doc, attribs={"name": self.remove_nodes_re}
+        )
+        parsers.remove(naughty_list)
+
+        # Navigation, menus, headers, footers, etc.
+        bad_tags = ["aside", "nav", "noscript", "menu"]
+        naughty_list = parsers.get_elements_by_tagslist(doc, bad_tags)
+        parsers.remove(naughty_list)
+
         return doc
 
-    def remove_nodes_regex(self, doc, pattern):
-        for selector in ["id", "class"]:
-            reg = "//*[re:test(@%s, '%s', 'i')]" % (selector, pattern)
-            naughty_list = self.parser.xpath_re(doc, reg)
-            for node in naughty_list:
-                self.parser.remove(node)
+    def remove_nodes_regex(self, doc: HtmlElement, pattern: str) -> HtmlElement:
+        """Removes HTML nodes from the given document that match the specified
+        regex pattern.
+
+        Args:
+            doc (HtmlElement): The HTML document to remove nodes from.
+            pattern (str): The regex pattern to match against the node attributes.
+
+        Returns:
+            HtmlElement: The modified HTML document with the matched nodes removed.
+        """
+        naughty_list = parsers.get_tags_regex(doc, attribs={"id": pattern})
+        naughty_list += parsers.get_tags_regex(doc, attribs={"class": pattern})
+
+        parsers.remove(naughty_list)
+
         return doc
 
-    def clean_para_spans(self, doc):
-        spans = self.parser.css_select(doc, "p span")
-        for item in spans:
-            self.parser.drop_tag(item)
+    def clean_para_spans(self, doc: HtmlElement) -> HtmlElement:
+        """Removes span tags within paragraph tags from the given HTML document.
+
+        Args:
+            doc (HtmlElement): The HTML document to clean.
+
+        Returns:
+            HtmlElement: The cleaned HTML document.
+        """
+        spans = doc.xpath(".//p/span")
+        parsers.drop_tags(spans)
         return doc
 
-    def get_flushed_buffer(self, replacement_text, doc):
-        return self.parser.textToPara(replacement_text)
+    def reduce_article(self, doc: HtmlElement) -> HtmlElement:
+        """Reduces the article by removing unnecessary tags from the
+        given HTML document. Keeps only tags that might contain the article and
+        its images.
 
-    def replace_walk_left_right(self, kid, kid_text, replacement_text, nodes_to_remove):
-        kid_text_node = kid
-        replace_text = self.tablines_replacements.replaceAll(kid_text)
-        if len(replace_text) > 1:
-            prev_node = self.parser.previousSibling(kid_text_node)
-            while (
-                prev_node is not None
-                and self.parser.getTag(prev_node) == "a"
-                and self.parser.getAttribute(prev_node, "grv-usedalready") != "yes"
-            ):
-                outer = " " + self.parser.outerHtml(prev_node) + " "
-                replacement_text.append(outer)
-                nodes_to_remove.append(prev_node)
-                self.parser.setAttribute(prev_node, attr="grv-usedalready", value="yes")
-                prev_node = self.parser.previousSibling(prev_node)
+        Args:
+            doc (HtmlElement): The HTML document to be reduced.
 
-            replacement_text.append(replace_text)
-            next_node = self.parser.nextSibling(kid_text_node)
-            while (
-                next_node is not None
-                and self.parser.getTag(next_node) == "a"
-                and self.parser.getAttribute(next_node, "grv-usedalready") != "yes"
-            ):
-                outer = " " + self.parser.outerHtml(next_node) + " "
-                replacement_text.append(outer)
-                nodes_to_remove.append(next_node)
-                self.parser.setAttribute(next_node, attr="grv-usedalready", value="yes")
-                next_node = self.parser.nextSibling(next_node)
+        Returns:
+            HtmlElement: The reduced HTML document.
+        """
+        body_tag = parsers.get_tags(doc, tag="body")
+        if not body_tag:
+            return doc
 
-    def get_replacement_nodes(self, doc, div):
-        replacement_text = []
-        nodes_to_return = []
-        nodes_to_remove = []
-        kids = self.parser.childNodesWithText(div)
-        for kid in kids:
-            # The node is a <p> and already has some replacement text
-            if self.parser.getTag(kid) == "p" and len(replacement_text) > 0:
-                new_node = self.get_flushed_buffer("".join(replacement_text), doc)
-                nodes_to_return.append(new_node)
-                replacement_text = []
-                nodes_to_return.append(kid)
-            # The node is a text node
-            elif self.parser.isTextNode(kid):
-                kid_text = self.parser.getText(kid)
-                self.replace_walk_left_right(
-                    kid, kid_text, replacement_text, nodes_to_remove
-                )
-            else:
-                nodes_to_return.append(kid)
+        for item in body_tag[0].iter():
+            if item.tag not in [
+                "p",
+                "br",
+                "img",
+                "h1",
+                "h2",
+                "h3",
+                "h4",
+                "h5",
+                "h6",
+                "ul",
+                "body",
+                "article",
+                "section",
+            ]:
+                if item.text is None and item.tail is None:
+                    item.drop_tag()
 
-        # flush out anything still remaining
-        if len(replacement_text) > 0:
-            new_node = self.get_flushed_buffer("".join(replacement_text), doc)
-            nodes_to_return.append(new_node)
-            replacement_text = []
-
-        for n in nodes_to_remove:
-            self.parser.remove(n)
-
-        return nodes_to_return
-
-    def replace_with_para(self, doc, div):
-        self.parser.replaceTag(div, "p")
-
-    def div_to_para(self, doc, dom_type):
-        bad_divs = 0
-        else_divs = 0
-        divs = self.parser.getElementsByTag(doc, tag=dom_type)
-        tags = ["a", "blockquote", "dl", "div", "img", "ol", "p", "pre", "table", "ul"]
-        for div in divs:
-            items = self.parser.getElementsByTags(div, tags)
-            if div is not None and len(items) == 0:
-                self.replace_with_para(doc, div)
-                bad_divs += 1
-            elif div is not None:
-                replace_nodes = self.get_replacement_nodes(doc, div)
-                replace_nodes = [n for n in replace_nodes if n is not None]
-                attrib = copy.deepcopy(div.attrib)
-                div.clear()
-                for i, node in enumerate(replace_nodes):
-                    div.insert(i, node)
-                for name, value in attrib.items():
-                    div.set(name, value)
-                else_divs += 1
         return doc
