@@ -1,16 +1,36 @@
 import copy
+from functools import partial
+import re
+from statistics import mean
+from typing import Optional
 import lxml
+from newspaper.configuration import Configuration
 import newspaper.extractors.defines as defines
+import newspaper.parsers as parsers
+from newspaper.text import StopWords
+
+score_weights = {
+    "bottom_negativescore_nodes": 0.25,
+    "boost_score": 30,
+    "parent_node": 1.0,
+    "parent_parent_node": 0.4,
+    "node_count_threshold": 15,
+    "negative_score_threshold": 40,
+    "negative_score_boost": 5.0,
+    "boost_max_steps_from_node": 3,
+    "boost_min_stopword_count": 5,
+}
+
+get_stop_words = partial(parsers.get_attribute, attr="stop_words", type_=int, default=0)
+get_word_count = partial(parsers.get_attribute, attr="word_count", type_=int, default=0)
 
 
 class ArticleBodyExtractor:
-    def __init__(self, config):
+    def __init__(self, config: Configuration):
         self.config = config
         self.top_node = None
         self.top_node_complemented = None
-        self.parser = self.config.get_parser()
-        self.stopwords_class = config.stopwords_class
-        self.language = config.language
+        self.stopwords: Optional[StopWords] = None
 
     def parse(self, doc: lxml.html.Element):
         """_summary_
@@ -18,76 +38,137 @@ class ArticleBodyExtractor:
         Args:
             doc (lxml.html.Element): _description_
         """
+        self.stopwords = StopWords(self.config.language)
         self.top_node = self.calculate_best_node(doc)
         self.top_node_complemented = self.complement_with_siblings(self.top_node)
 
     def calculate_best_node(self, doc):
         top_node = None
-        nodes_to_check = self.nodes_to_check(doc)
         self.boost_highly_likely_nodes(doc)
-        starting_boost = float(1.0)
-        cnt = 0
-        i = 0
+
         parent_nodes = []
-        nodes_with_text = []
+        nodes_with_text = self.compute_features(doc)
 
-        for node in nodes_to_check:
-            text_node = self.parser.getText(node)
-            word_stats = self.stopwords_class(
-                language=self.language
-            ).get_stopword_count(text_node)
-            high_link_density = self.is_highlink_density(node)
-            if word_stats.stop_word_count > 2 and not high_link_density:
-                nodes_with_text.append(node)
+        # process the tree from bottom up. farthest nodes first
+        nodes_with_text.sort(
+            key=lambda node: parsers.get_attribute(
+                node, "node_level", type_=int, default=0
+            ),
+            reverse=True,
+        )
 
-        nodes_number = len(nodes_with_text)
-        negative_scoring = 0
-        bottom_negativescore_nodes = float(nodes_number) * 0.25
-
-        for node in nodes_with_text:
-            boost_score = float(0)
-            # boost
-            if self.is_boostable(node):
-                if cnt >= 0:
-                    boost_score = float((1.0 / starting_boost) * 50)
-                    starting_boost += 1
-            # nodes_number
-            if nodes_number > 15:
-                if (nodes_number - i) <= bottom_negativescore_nodes:
-                    booster = float(bottom_negativescore_nodes - (nodes_number - i))
-                    boost_score = float(-pow(booster, float(2)))
-                    negscore = abs(boost_score) + negative_scoring
-                    if negscore > 40:
-                        boost_score = float(5)
-
-            text_node = self.parser.getText(node)
-            word_stats = self.stopwords_class(
-                language=self.language
-            ).get_stopword_count(text_node)
-            upscore = int(word_stats.stop_word_count + boost_score)
-
-            parent_node = self.parser.getParent(node)
-            self.update_score(parent_node, upscore)
-            self.update_node_count(parent_node, 1)
-
-            if parent_node not in parent_nodes:
-                parent_nodes.append(parent_node)
-
-            # Parent of parent node
-            parent_parent_node = self.parser.getParent(parent_node)
-            if parent_parent_node is not None:
-                self.update_node_count(parent_parent_node, 1)
-                self.update_score(parent_parent_node, upscore / 2)
-                if parent_parent_node not in parent_nodes:
-                    parent_nodes.append(parent_parent_node)
-            cnt += 1
-            i += 1
+        parent_nodes = self.compute_gravity_scores(nodes_with_text)
 
         if parent_nodes:
-            parent_nodes.sort(key=self.get_score, reverse=True)
+            parent_nodes.sort(key=parsers.get_node_gravity_score, reverse=True)
             top_node = parent_nodes[0]
 
         return top_node
+
+    def compute_gravity_scores(self, nodes_with_text):
+        """Computes the gravity score for each node in the list.
+        And propagate the score to its parents and grandparents.
+
+        Args:
+            nodes_with_text (list): list of candidate nodes that have meaningful text
+
+        Returns:
+            list: list of nodes with gravity score
+        """
+        parent_nodes = []
+
+        boost_discount = 1
+
+        nodes_count = len(nodes_with_text)
+        negative_scoring = 0
+
+        bottom_negativescore_nodes = (
+            nodes_count * score_weights["bottom_negativescore_nodes"]
+        )
+
+        for i, node in enumerate(nodes_with_text):
+            boost_score = 0
+            # boost score decays with distance from the top node
+            if self.is_boostable(node):
+                boost_score = score_weights["boost_score"] / boost_discount
+                boost_discount += 1
+
+            # nodes_number
+            if nodes_count > score_weights["node_count_threshold"]:
+                # higher number of possible top nodes
+                dist_from_end = float(nodes_count - i)
+                if dist_from_end <= bottom_negativescore_nodes:
+                    booster = bottom_negativescore_nodes - dist_from_end
+                    boost_score = -(booster**2)
+                    negscore = abs(boost_score) + negative_scoring
+                    if negscore > score_weights["negative_score_threshold"]:
+                        boost_score = score_weights["negative_score_boost"]
+
+            stop_word_count = get_stop_words(node)
+
+            upscore = stop_word_count + boost_score
+
+            parent_node = node.getparent()
+
+            self.update_score(parent_node, upscore)
+            self.update_node_count(parent_node, 1)
+
+            parent_nodes.append(parent_node)
+
+            # Parent of parent node
+            parent_parent_node = (
+                parent_node.getparent() if parent_node is not None else None
+            )
+
+            self.update_node_count(parent_parent_node, 1)
+            self.update_score(
+                parent_parent_node, upscore * score_weights["parent_parent_node"]
+            )
+
+            parent_nodes.append(parent_parent_node)
+
+        parent_nodes = [x for x in set(parent_nodes) if x is not None]
+
+        return parent_nodes
+
+    def compute_features(self, doc):
+        candidates = []
+        nodes_to_check = self.nodes_to_check(doc)
+        nodes_to_check.sort(key=parsers.get_level, reverse=True)
+
+        for node in nodes_to_check:
+            # exclude nodes that are in this list
+
+            text_content = parsers.get_text(node)
+            if not text_content:
+                continue
+
+            word_stats = self.stopwords.get_stopword_count(text_content)
+            high_link_density = parsers.is_highlink_density(node, self.config.language)
+
+            children_word_stats = [
+                (get_stop_words(child), get_word_count(child))
+                for child in node.xpath(".//*[@stop_words>0]")
+            ]
+            children_word_stats = (
+                sum([x[0] for x in children_word_stats]),
+                sum([x[1] for x in children_word_stats]),
+            )
+            parsers.set_attribute(
+                node, "stop_words", word_stats.stop_word_count - children_word_stats[0]
+            )
+            parsers.set_attribute(
+                node, "word_count", word_stats.word_count - children_word_stats[1]
+            )
+            parsers.set_attribute(
+                node, "is_highlink_density", 1 if high_link_density else 0
+            )
+            parsers.set_attribute(node, "node_level", parsers.get_level(node))
+
+            if word_stats.stop_word_count > 2 and not high_link_density:
+                candidates.append(node)
+
+        return candidates
 
     def nodes_to_check(self, doc):
         """Returns a list of nodes we want to search
@@ -97,13 +178,44 @@ class ArticleBodyExtractor:
         for tag in ["p", "pre", "td", "article", "div"]:
             if tag == "div":
                 items = []
-                for id_ in ["article-body", "article", "story", "article-content"]:
-                    items += self.parser.getElementsByTag(
-                        doc, tag=tag, attr="id", value=id_
+                for attr in [
+                    "articlebody",
+                    "article",
+                    "story",
+                ]:
+                    items += parsers.get_tags(
+                        doc,
+                        tag=tag,
+                        attribs={"id": attr},
+                        attribs_match="word",
+                        ignore_dashes=True,
                     )
+                    items += parsers.get_tags(
+                        doc,
+                        tag=tag,
+                        attribs={"class": attr},
+                        attribs_match="word",
+                        ignore_dashes=True,
+                    )
+                for class_ in ["paragraph"]:
+                    items += parsers.get_tags_regex(
+                        doc, tag=tag, attribs={"class": class_}
+                    )
+                if len(items) == 0 and len(nodes_to_check) < 5:
+                    items = parsers.get_tags(doc, tag=tag)
+                items = set(items)  # remove duplicates
             else:
-                items = self.parser.getElementsByTag(doc, tag=tag)
+                items = parsers.get_tags(doc, tag=tag)
             nodes_to_check += items
+
+        # Do not miss some Article Bodies or Article Sections
+        # for itemprop in ['articleBody', 'articlebody', 'articleText',
+        #                       'articleSection']:
+        #     items = [item  for item in parsers.get_tags(
+        #         doc, attribs={"itemprop": itemprop}, attribs_match="word"
+        #     ) if item not in nodes_to_check]
+        #     nodes_to_check.extend(items)
+
         return nodes_to_check
 
     def is_boostable(self, node):
@@ -113,54 +225,18 @@ class ArticleBodyExtractor:
         paragraphs so we'll want to make sure that the next sibling is a
         paragraph and has at least some substantial weight to it.
         """
-        para = "p"
-        steps_away = 0
-        minimum_stopword_count = 5
-        max_stepsaway_from_node = 3
+        max_stepsaway_from_node = score_weights["boost_max_steps_from_node"]
 
         nodes = self.walk_siblings(node)
-        for current_node in nodes:
-            # <p>
-            current_node_tag = self.parser.getTag(current_node)
-            if current_node_tag == para:
-                if steps_away >= max_stepsaway_from_node:
-                    return False
-                paragraph_text = self.parser.getText(current_node)
-                word_stats = self.stopwords_class(
-                    language=self.language
-                ).get_stopword_count(paragraph_text)
-                if word_stats.stop_word_count > minimum_stopword_count:
-                    return True
-                steps_away += 1
+        for current_node in nodes[:max_stepsaway_from_node]:
+            if current_node.tag != node.tag:
+                continue
+            stop_word_count = parsers.get_attribute(
+                current_node, "stop_words", type_=int, default=0
+            )
+            if stop_word_count > score_weights["boost_min_stopword_count"]:
+                return True
         return False
-
-    def is_highlink_density(self, e):
-        """Checks the density of links within a node, if there is a high
-        link to text ratio, then the text is less likely to be relevant
-        """
-        links = self.parser.getElementsByTag(e, tag="a")
-        if not links:
-            return False
-
-        text = self.parser.getText(e)
-        words = [word for word in text.split() if word.isalnum()]
-        if not words:
-            return True
-        words_number = float(len(words))
-        sb = []
-        for link in links:
-            sb.append(self.parser.getText(link))
-
-        link_text = "".join(sb)
-        link_words = link_text.split()
-        num_link_words = float(len(link_words))
-        num_links = float(len(links))
-        link_divisor = float(num_link_words / words_number)
-        score = float(link_divisor * num_links)
-        if score >= 1.0:
-            return True
-        return False
-        # return True if score > 1.0 else False
 
     def boost_highly_likely_nodes(self, doc: lxml.html.Element):
         """Set a bias score for all nodes under most likely
@@ -172,14 +248,17 @@ class ArticleBodyExtractor:
         """
         candidates = []
         for tag in ["p", "pre", "td", "article", "div"]:
-            candidates.extend(self.parser.getElementsByTag(doc, tag=tag))
+            candidates.extend(parsers.get_tags(doc, tag=tag))
 
         for e in candidates:
-            if self.is_highly_likly(e):
-                for child in e.iterdescendants():
-                    self.update_score(child, 25)  # TODO: find an optimum value
+            boost = self.is_highly_likely(e)
+            if boost > 0:
+                self.update_score(e, boost * score_weights["parent_node"])
+                # for child in e.iterdescendants():
+                # TODO: find an optimum value
+                #     self.update_score(child, boost // 10)
 
-    def is_highly_likly(self, node: lxml.html.Element) -> bool:
+    def is_highly_likely(self, node: lxml.html.Element) -> int:
         """Checks if the node is a well known tag + attributes combination
         for article body containers. This way we can deliver even small
         article bodies with high link density
@@ -197,85 +276,96 @@ class ArticleBodyExtractor:
             if node.tag != tag_dict.get("tag", node.tag):
                 return False
             for k, v in tag_dict.items():
-                if k == "tag":
+                if k in ["tag", "score_boost"]:
                     continue
-                if node.get(k) != v:
+                if v.startswith("re:"):
+                    v = v[3:]
+                    if not re.search(v, node.get(k, ""), re.IGNORECASE):
+                        return False
+                elif node.get(k, "").lower() != v.lower():
                     return False
             return True
 
+        scores = []
         for tag in defines.ARTICLE_BODY_TAGS:
             if is_tag_match(node, tag):
-                return True
+                scores.append(tag["score_boost"])
+        if scores:
+            return max(scores)
 
-        return False
+        return 0
 
     def update_score(self, node, add_to_score):
         """Adds a score to the gravityScore Attribute we put on divs
         we'll get the current score then add the score we're passing
         in to the current.
         """
-        current_score = 0
-        score_string = self.parser.getAttribute(node, "gravityScore")
-        if score_string:
-            current_score = float(score_string)
-
-        new_score = current_score + add_to_score
-        self.parser.setAttribute(node, "gravityScore", str(new_score))
+        if node is None:
+            return
+        new_score = parsers.get_node_gravity_score(node) + add_to_score
+        parsers.set_attribute(node, "gravityScore", str(new_score))
 
     def update_node_count(self, node, add_to_count):
         """Stores how many decent nodes are under a parent node"""
-        current_score = 0
-        count_string = self.parser.getAttribute(node, "gravityNodes")
-        if count_string:
-            current_score = int(count_string)
-
-        new_score = current_score + add_to_count
-        self.parser.setAttribute(node, "gravityNodes", str(new_score))
+        if node is None:
+            return
+        new_count = float(node.get("gravityNodes", 0)) + add_to_count
+        parsers.set_attribute(node, "gravityNodes", str(new_count))
 
     def add_siblings(self, top_node):
         res_node = copy.deepcopy(top_node)
-        baseline_score_siblings_para = self.get_siblings_score(top_node)
+        baseline_score = self.get_normalized_score(top_node)
         results = self.walk_siblings(top_node)
         for current_node in results:
-            ps = self.get_siblings_content(current_node, baseline_score_siblings_para)
+            ps = self.get_plausible_content(current_node, baseline_score)
             for p in ps:
                 res_node.insert(0, p)
         return res_node
 
-    def get_siblings_content(self, current_sibling, baseline_score_siblings_para):
-        """Adds any siblings that may have a decent score to this node"""
-        if current_sibling.tag == "p" and len(self.parser.getText(current_sibling)) > 0:
-            e0 = current_sibling
-            if e0.tail:
-                e0 = copy.deepcopy(e0)
-                e0.tail = ""
-            return [e0]
-        else:
-            potential_paragraphs = self.parser.getElementsByTag(
-                current_sibling, tag="p"
-            )
-            if potential_paragraphs is None:
-                return None
-            else:
-                ps = []
-                for first_paragraph in potential_paragraphs:
-                    text = self.parser.getText(first_paragraph)
-                    if len(text) > 0:
-                        word_stats = self.stopwords_class(
-                            language=self.language
-                        ).get_stopword_count(text)
-                        paragraph_score = word_stats.stop_word_count
-                        sibling_baseline_score = float(0.30)
-                        high_link_density = self.is_highlink_density(first_paragraph)
-                        score = float(
-                            baseline_score_siblings_para * sibling_baseline_score
-                        )
-                        if score < paragraph_score and not high_link_density:
-                            p = self.parser.createElement(tag="p", text=text, tail=None)
-                            ps.append(p)
-                return ps
+    def get_plausible_content(self, node, baseline_score, score_weight=0.3):
+        """Create list of (off-the-tree) paragraphs from a node (most likely a
+        sibling of the top node) that are plausible to be part of the
+        article body. We use the stop word count as a score of plausibility.
+        We accept only paragraphs with a score higher than the baseline score
+        weighted by the score_weight parameter.
+        The baseline score is a normalized score of the top node.
+        """
+        if isinstance(
+            node, (lxml.etree.CommentBase, lxml.etree.EntityBase, lxml.etree.PIBase)
+        ):
+            return []
 
-    def get_siblings_score(self, top_node):
+        if (
+            node.tag == "p"
+            and node.text
+            and not parsers.is_highlink_density(node, self.config.language)
+        ):
+            element = copy.deepcopy(node)
+            element.tail = ""
+            return [element]
+
+        paragraphs = parsers.get_tags(node, tag="p")
+        result = []
+        if not paragraphs:
+            return result
+
+        for paragraph in paragraphs:
+            stop_word_count = parsers.get_attribute(
+                paragraph, "stop_words", type_=int, default=0
+            )
+            if stop_word_count <= 0:
+                continue
+            if parsers.is_highlink_density(paragraph, self.config.language):
+                continue
+
+            if stop_word_count > baseline_score * score_weight:
+                text = parsers.get_text(paragraph)
+                element = parsers.create_element(tag="p", text=text)
+                result.append(element)
+
+        return result
+
+    def get_normalized_score(self, top_node):
         """We could have long articles that have tons of paragraphs
         so if we tried to calculate the base score against
         the total text score of those paragraphs it would be unfair.
@@ -284,38 +374,17 @@ class ArticleBodyExtractor:
         For example if our total score of 10 paragraphs was 1000
         but each had an average value of 100 then 100 should be our base.
         """
-        base = 100000
-        paragraphs_number = 0
-        paragraphs_score = 0
-        nodes_to_check = self.parser.getElementsByTag(top_node, tag="p")
 
-        for node in nodes_to_check:
-            text_node = self.parser.getText(node)
-            word_stats = self.stopwords_class(
-                language=self.language
-            ).get_stopword_count(text_node)
-            high_link_density = self.is_highlink_density(node)
-            if word_stats.stop_word_count > 2 and not high_link_density:
-                paragraphs_number += 1
-                paragraphs_score += word_stats.stop_word_count
+        nodes_to_check = parsers.get_tags(top_node, tag="p")
 
-        if paragraphs_number > 0:
-            base = paragraphs_score / paragraphs_number
+        scores = [parsers.get_node_gravity_score(node) for node in nodes_to_check]
+        scores = [score for score in scores if score > 0]  # filter out 0 scores
 
-        return base
+        return mean(scores) if scores else float("inf")
 
     def walk_siblings(self, node):
-        return self.parser.previousSiblings(node)
-
-    def get_node_gravity_score(self, node):
-        gravity_score = self.parser.getAttribute(node, "gravityScore")
-        if not gravity_score:
-            return None
-        return float(gravity_score)
-
-    def get_score(self, node):
-        """Returns the gravityScore as an integer from this node"""
-        return self.get_node_gravity_score(node) or 0
+        """returns preceding siblings in reverse order (nearest sibling is first)"""
+        return [n for n in node.itersiblings(preceding=True)]
 
     def complement_with_siblings(self, node: lxml.html.Element) -> lxml.html.Element:
         """Adds surrounding relevant siblings to the top node.
@@ -329,12 +398,52 @@ class ArticleBodyExtractor:
         """
         if node is None:
             return node
-        node_complemented = self.add_siblings(
-            node
-        )  # TODO: test if there is a problem with siblings AFTER the top node
-        for e in self.parser.getChildren(node_complemented):
-            e_tag = self.parser.getTag(e)
-            if e_tag != "p":
-                if self.is_highlink_density(e):
-                    self.parser.remove(e)
+        # TODO: test if there is a problem with siblings AFTER the top node
+        # node_complemented = self.add_siblings(node)
+
+        node_complemented = self.add_same_level_candidates(node)
+
+        # for e in node_complemented.getchildren():
+        #     if e.tag != "p":
+        #         if parsers.is_highlink_density(e):
+        #             parsers.remove(e)
         return node_complemented
+
+    def add_same_level_candidates(self, node):
+        """Adds any siblings that may have a decent score to this node"""
+        tree = node.getroottree()
+
+        node_level = parsers.get_level(node)
+        # base_score = self.get_normalized_score(node)
+        base_score = parsers.get_node_gravity_score(node)
+
+        candidates = parsers.get_nodes_at_level(tree.getroot(), node_level)
+
+        assert node in candidates, "node not in candidates"
+
+        # Create the new tree as a HTML document. Otherwise it will
+        # be created as Element tree and cleaners won't work
+        new_tree = parsers.fromstring("<html><body></body></html>")
+        new_node = new_tree.find("body")
+
+        for n in candidates:
+            if n == node:
+                new_node.append(copy.deepcopy(node))
+                continue
+
+            # avoid adding nodes that do not resemble the top node
+            if n.tag != node.tag:
+                continue
+
+            score = parsers.get_node_gravity_score(n)
+
+            if score > base_score * 0.3 and not parsers.is_highlink_density(
+                n, self.config.language
+            ):
+                new_node.append(copy.deepcopy(n))
+                continue
+
+            # content_items = self.get_plausible_content(n, base_score)
+            # new_node.extend(content_items)
+
+        return new_node
