@@ -5,15 +5,15 @@ from typing import Optional
 import lxml
 from newspaper import urls
 from newspaper.configuration import Configuration
+import newspaper.parsers as parsers
 from dateutil.parser import parse as date_parser
 
-from newspaper.extractors.defines import PUBLISH_DATE_TAGS
+from newspaper.extractors.defines import PUBLISH_DATE_META_INFO, PUBLISH_DATE_TAGS
 
 
 class PubdateExtractor:
     def __init__(self, config: Configuration) -> None:
         self.config = config
-        self.parser = config.get_parser()
         self.pubdate: Optional[datetime] = None
 
     def parse(self, article_url: str, doc: lxml.html.Element) -> Optional[datetime]:
@@ -39,13 +39,13 @@ class PubdateExtractor:
         date_matches = []
         date_match = re.search(urls.STRICT_DATE_REGEX, article_url)
         if date_match:
-            date_str = date_match.group(0)
-            datetime_obj = parse_date_str(date_str)
+            date_match_str = date_match.group(0)
+            datetime_obj = parse_date_str(date_match_str)
             if datetime_obj:
                 date_matches.append((datetime_obj, 10))  # date and matchscore
 
         # yoast seo structured data or json-ld
-        json_ld_scripts = self.parser.get_ld_json_object(doc)
+        json_ld_scripts = parsers.get_ld_json_object(doc)
 
         for script_tag in json_ld_scripts:
             if "@graph" in script_tag:
@@ -54,6 +54,8 @@ class PubdateExtractor:
                     if not isinstance(item, dict):
                         continue
                     date_str = item.get("datePublished")
+                    if date_str is None:
+                        continue
                     datetime_obj = parse_date_str(date_str)
                     if datetime_obj:
                         date_matches.append((datetime_obj, 10))
@@ -66,7 +68,7 @@ class PubdateExtractor:
                             date_matches.append((datetime_obj, 9))
 
         # get <time> tags
-        for item in self.parser.getElementsByTag(doc, tag="time"):
+        for item in parsers.get_tags(doc, tag="time"):
             if item.get("datetime"):
                 date_str = item.get("datetime")
                 datetime_obj = parse_date_str(date_str)
@@ -77,24 +79,37 @@ class PubdateExtractor:
                         )  # Boost if it has the word published or on
                     else:
                         date_matches.append((datetime_obj, 5))
+        candidates = []
 
+        for known_meta_info in PUBLISH_DATE_META_INFO:
+            candidates.extend(parsers.get_metatags(doc, value=known_meta_info))
+        candidates = [(x, "content") for x in candidates]  # property that contains
+        # the date
+        # is always 'content'
         for known_meta_tag in PUBLISH_DATE_TAGS:
-            meta_tags = self.parser.getElementsByTag(
-                doc, attr=known_meta_tag["attribute"], value=known_meta_tag["value"]
+            candidates.extend(
+                [
+                    (x, known_meta_tag["content"])
+                    for x in parsers.get_elements_by_attribs(
+                        doc,
+                        attribs={known_meta_tag["attribute"]: known_meta_tag["value"]},
+                    )
+                ]
             )
-            for meta_tag in meta_tags:
-                date_str = self.parser.getAttribute(meta_tag, known_meta_tag["content"])
-                datetime_obj = parse_date_str(date_str)
-                if datetime_obj:
-                    score = 6
-                    if meta_tag.attrib.get("name") == known_meta_tag["value"]:
-                        score += 2
-                    days_diff = (datetime.now().date() - datetime_obj.date()).days
-                    if days_diff < 0:  # articles from the future
-                        score -= 2
-                    elif days_diff > 25 * 365:  # very old articles
-                        score -= 1
-                    date_matches.append((datetime_obj, score))
+
+        for meta_tag, content_attr in candidates:
+            date_str = parsers.get_attribute(meta_tag, content_attr)
+            datetime_obj = parse_date_str(date_str)
+            if datetime_obj:
+                score = 6
+                if meta_tag.tag.lower() == "meta":
+                    score += 1  # Boost meta tags
+                days_diff = (datetime.now().date() - datetime_obj.date()).days
+                if days_diff < 0:  # articles from the future
+                    score -= 2
+                elif days_diff > 25 * 365:  # very old articles
+                    score -= 1
+                date_matches.append((datetime_obj, score))
 
         date_matches.sort(key=lambda x: x[1], reverse=True)
         self.pubdate = date_matches[0][0] if date_matches else None
